requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg[binary]==3.3.2

# Email
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
hypothesis==6.92.1
pytest-postgresql==5.0.0
freezegun==1.4.0
responses==0.24.1

# Security
Flask-Bcrypt==1.0.1
